requirements.txt
nose==1.3.7
python-chess==0.23.9
